standardize track linking
